refactor: query hook을 entites에 따라 분리합니다
