+++ apps/gomterview-com/src/components/common/QuestionSelectionBox/QuestionTabPannel/QuestionTabPanelItem.tsx
@@ -1,14 +1,20 @@
 import { theme } from '@gomterview/_theme';
 import { css } from '@emotion/react';
 import { useState } from 'react';
-import { Button, Icon, Toggle, Typography } from 'gomterview-design-system';
+import {
+  Button,
+  EmptySuspense,
+  Icon,
+  Toggle,
+  Typography,
+  useTabs,
+} from 'gomterview-design-system';
 import { WorkbookTitleListResDto } from '@/types/workbook';
 import { ExcludeArray } from '@/types/utils';
 import QuestionTabPanelHeader from '@common/QuestionSelectionBox/QuestionTabPannel/QuestionTabPanelHeader';
-import { useTabs, EmptySuspense } from 'gomterview-design-system';
 import useBreakpoint from '@hooks/useBreakPoint';
 import QuestionTabPanelEditHeader from './QuestionTabPanelEditHeader';
-import useWorkbookQuery from '@hooks/apis/queries/useWorkbookQuery';
+import useWorkbookQuery from '@/entities/workbook/model/queries/useWorkbookQuery';
 import QuestionTabPanelBlank from '@common/QuestionSelectionBox/QuestionTabPannel/QuestionTabPanelBlank';
 import QuestionAccordionList from '@common/QuestionSelectionBox/QuestionTabPannel/QuestionAccordionList';
