+++ apps/gomterview-com/src/components/workbookDetailPage/NewWorkbookListButton.tsx
@@ -1,8 +1,8 @@
 import { WorkbookEntity } from '@/types/workbook';
 import { css } from '@emotion/react';
 import { Icon, Typography } from 'gomterview-design-system';
-import useQuestionCopyMutation from '@hooks/apis/mutations/useQuestionCopyMutation';
-import useWorkbookPostMutation from '@hooks/apis/mutations/useWorkbookPostMutation';
+import useQuestionCopyMutation from '@/entities/question/model/mutations/useQuestionCopyMutation';
+import useWorkbookPostMutation from '@/entities/workbook/model/mutations/useWorkbookPostMutation';
 import { toast } from '@gomterview/toast';
 
 const NewWorkbookListButton = ({
