+++ apps/gomterview-com/src/components/workbookPage/CategoryMenu.tsx
@@ -1,8 +1,7 @@
 import { css } from '@emotion/react';
-import { Box, SelectionBox, Typography } from 'gomterview-design-system';
+import { Box, SelectionBox, Tabs, Typography } from 'gomterview-design-system';
 import { ResponsiveMenu } from '@common/index';
-import { Tabs } from 'gomterview-design-system';
-import useCategoryQuery from '@hooks/apis/queries/useCategoryQuery';
+import useCategoryQuery from '@/entities/category/model/queries/useCategoryQuery';
 import useBreakpoint from '@hooks/useBreakPoint';
 
 type CategoryMenuProps = {
