+++ apps/gomterview-com/src/entities/question/model/mutations/useDeleteQuestionMutation.tsx
