+++ apps/gomterview-com/src/entities/question/model/mutations/useQuestionAnswerMutation.ts
@@ -1,5 +1,5 @@
 import { postAnswer } from '@/apis/answer';
-import { QUERY_KEY } from '@/constants/queryKey';
+import { QUERY_KEY } from '@constants/queryKey';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 /**
