+++ apps/gomterview-com/src/entities/workbook/model/queries/useWorkbookQuery.ts
@@ -1,10 +1,10 @@
 import {
-  UseSuspenseQueryResult,
   useSuspenseQuery,
+  UseSuspenseQueryResult,
 } from '@tanstack/react-query';
 import { QUERY_KEY } from '@constants/queryKey';
 import { getWorkbookById } from '@/apis/workbook';
-import useCategoryQuery from './useCategoryQuery';
+import useCategoryQuery from '../../../category/model/queries/useCategoryQuery';
 import { WorkbookEntity } from '@/types/workbook';
 import useUserInfo from '@hooks/useUserInfo';
