+++ apps/gomterview-com/src/hooks/useQuestionAdd.ts
@@ -1,5 +1,5 @@
 import { useQueryClient } from '@tanstack/react-query';
-import useQuestionMutation from './apis/mutations/useQuestionMutation';
+import useQuestionMutation from '../entities/question/model/mutations/useQuestionMutation';
 import useUserInfo from './useUserInfo';
 import { Question } from '@/types/question';
 import { QUERY_KEY } from '@/constants/queryKey';
