+++ apps/gomterview-com/src/hooks/useUploadToIdrive.ts
@@ -1,7 +1,7 @@
 import { SelectedQuestion } from '@/atoms/interviewSetting';
-import useGetPreSignedUrlMutation from '@/hooks/apis/mutations/useGetPreSignedUrlMutation';
+import useGetPreSignedUrlMutation from '@/entities/video/model/mutations/useGetPreSignedUrlMutation';
 import { putBlobDataToIdrive } from '@/apis/idrive';
-import useAddVideoMutation from '@/hooks/apis/mutations/useAddVideoMutation';
+import useAddVideoMutation from '@/entities/video/model/mutations/useAddVideoMutation';
 import { toast } from '@gomterview/toast';
 import { EncodingWebmToMp4, getThumbnailBlob } from '@/utils/record';
 import { IDRIVE_URL } from '@constants/api';
