+++ apps/gomterview-com/src/page/interviewVideoPage/index.tsx
@@ -1,7 +1,7 @@
 import { LoadingBounce, VideoList } from '@common/index';
 import {
-  InterviewVideoPageLayout,
   DefaultVideoPlayer,
+  InterviewVideoPageLayout,
 } from '@components/interviewVideoPage';
 import { Suspense, useEffect, useState } from 'react';
 import { Navigate, useParams } from 'react-router-dom';
@@ -9,7 +9,7 @@ import { CenterLayout } from '@components/layout';
 import { PATH } from '@constants/path';
 import { ErrorBoundary } from 'react-error-boundary';
 import { isAxiosError } from 'axios';
-import useOnlyRelatedVideoQuery from '@hooks/apis/queries/useOnlyRelatedVideoListQuery';
+import useOnlyRelatedVideoQuery from '@/entities/video/model/queries/useOnlyRelatedVideoListQuery';
 import { Box, Typography } from 'gomterview-design-system';
 import { css } from '@emotion/react';
 import { theme } from '@gomterview/_theme';
