+++ packages/toast/src/ToastItem/ToastItem.tsx
@@ -2,11 +2,6 @@ import { useRef, useState } from 'react';
 import { ToastEvent, ToastProps } from '../type';
 import { css } from '@emotion/react';
 import { theme } from '@gomterview/_theme';
-import {
-  Box,
-  Icon,
-  Typography,
-} from '/Users/yhm/project/Gomterview-FE/packages/design-system/src/components';
 import { collapseToast } from '../collapseToast';
 import { eventManager } from '../eventManger';
 import {
@@ -20,6 +15,7 @@ import {
 } from '../styles/ToastAnimation.styles';
 import { TOAST_DEFAULT_POSITION } from '../constants';
 import ToastToggleButton from './ToastToggleButton';
+import { Box, Icon, Typography } from 'gomterview-design-system';
 
 const ToastItem: React.FC<ToastProps> = ({
   toastId,
